A - Frog 1
A - Frog 1

+++ EducationalDPContest/dp_a/dp_a_rec_ready.cpp
@@ -30,13 +30,13 @@ int n;
 int h[MAX];
 int dp[MAX];
 
-long long solve(int pos) {
+int solve(int pos) {
   if(pos > n) return INF;
   if(pos == n) return 0;
   if(dp[pos] != -1) return dp[pos];
 
-  long long cost1 = abs(h[pos + 1] - h[pos]);
-  long long cost2 = abs(h[pos + 2] - h[pos]);
+  int cost1 = abs(h[pos + 1] - h[pos]);
+  int cost2 = abs(h[pos + 2] - h[pos]);
 
   return dp[pos] = min(cost1 + solve(pos + 1), cost2 + solve(pos + 2));
 }
